+++ GLP Projekte/Shapes/trunk/src/Shapes/Rectangle.java
@@ -13,8 +13,8 @@ public class Rectangle extends Polygon {
 	private double lx;
 	private double ly;
 	private double phiGlobal = 0;
-	private Double cx;
-	private Double cy;
+	private double cx;
+	private double cy;
 
 	/**
 	 * Konstruktor fuer Objekte der class Rectangle erzeugt ein achsenparalleles
@@ -27,13 +27,13 @@ public class Rectangle extends Polygon {
 	 * @param ly
 	 *            - Länge der Seiten in Richtung y-Achse
 	 */
-	public Rectangle(Point bottomLeft, int lx, int ly) {
+	public Rectangle(Point bottomLeft, double lx, double ly) {
 		this.lx = lx;
 		this.ly = ly;
 		this.points.add(new Point(bottomLeft.getX(), bottomLeft.getY()));
-		cx = (double) (points.get(0).getX() + lx / 2);
-		cy = (double) (points.get(0).getY() + ly / 2);
-		this.points.add(new Point(cx.intValue(), cy.intValue()));
+		cx =  (points.get(0).getX() + lx / 2);
+		cy =  (points.get(0).getY() + ly / 2);
+		this.points.add(new Point(cx, cy));
 	}
 
 	/**
@@ -74,19 +74,6 @@ public class Rectangle extends Polygon {
 		 while (it.hasNext()) {
 		 it.next().rotate(center, phi);
 		 }
-		// Double X=this.getLx();
-		// Double Y=this.getLy();
-		// Iterator<Point> it2 = points.iterator();
-		// while (it2.hasNext()) {
-		// it2.next().move(-X.intValue()/4,-Y.intValue()/4);
-		// }
-		// Double X=points.get(0).getX() + lx / 2;
-		// Double Y=points.get(0).getY() + ly / 2;
-		// Point Berechnung = new Point (X.intValue(),Y.intValue());
-		// Berechnung.rotate(new Point(1+X.intValue(),1+Y.intValue()), phi);
-		// this.setLx(Berechnung.getX());
-		// this.setLy(Berechnung.getY());
-
 		this.setPhiGlobal(this.getPhiGlobal() + (-Math.toRadians(phi)));
 		this.draw();
 		return this;
